NLP/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
backoff==2.1.2
beautifulsoup4==4.11.1
blis==0.7.8
Bottleneck==1.3.5
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cycler==0.11.0
cymem==2.0.6
fastai==1.0.57
fastprogress==1.0.3
fonttools==4.34.4
frozenlist==1.3.1
google-api-core==2.8.2
google-auth==2.10.0
google-cloud==0.34.0
google-cloud-speech==2.15.0
google-cloud-texttospeech==2.12.0
googleapis-common-protos==1.56.4
gql==3.4.0
graphql-core==3.2.1
grpcio==1.47.0
grpcio-status==1.47.0
idna==3.3
inltk==0.9
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.5.2
multidict==6.0.2
murmurhash==1.0.7
numexpr==2.8.3
numpy==1.23.1
nvidia-ml-py3==7.352.0
packaging==21.3
pandas==1.4.3
pathy==0.6.2
Pillow==9.2.0
preshed==3.0.6
proto-plus==1.20.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.0
sentencepiece==0.1.97
six==1.16.0
sklearn==0.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.9
spacy-loggers==1.0.3
srsly==2.4.4
thinc==8.1.0
threadpoolctl==3.1.0
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
typer==0.4.2
typing==3.7.4.3
typing_extensions==4.3.0
urllib3==1.26.11
wasabi==0.10.1
yarl==1.8.1